myousica: added song deletion user interface
git-svn-id: file:///srv/svn/myousica/myousica/trunk@1237 1758279c-6b0d-4dd2-8212-40bfc25400ca

+++ lib/playable.rb
@@ -41,6 +41,7 @@ module Adelao
 
         def delete_sound_file
           return unless self.filename
+          return if RAILS_ENV == 'development'
 
           File.unlink absolute_filename(:stream)
           File.unlink absolute_filename(:waveform)
